Enemy shoot

+++ Space Invaders Project/Space Invaders Project/clock.cpp
@@ -44,11 +44,11 @@ CClock::Initialise()
 {
 	__int64 _TimerFrequency, _currTime;
 	QueryPerformanceFrequency((LARGE_INTEGER*)&_TimerFrequency);
-	m_SecondsPerCount = 1.0f / static_cast<float>(_TimerFrequency);
+	m_SecondsPerCount = 1.0 / static_cast<double>(_TimerFrequency);
 
 	QueryPerformanceCounter((LARGE_INTEGER*)&_currTime);
-	m_fCurrentTime = static_cast<float>(_currTime);
-	m_fLastTime = static_cast<float>(_currTime);
+	m_fCurrentTime = static_cast<double>(_currTime);
+	m_fLastTime = static_cast<double>(_currTime);
 
 	return (true);
 }
@@ -62,7 +62,7 @@ CClock::Process()
 
 	QueryPerformanceCounter((LARGE_INTEGER*)&currTime);
 
-	m_fCurrentTime = static_cast<float>(currTime);
+	m_fCurrentTime = static_cast<double>(currTime);
 
 	//Time difference between this frame and the previous frame
 	m_fDeltaTime = (m_fCurrentTime - m_fLastTime)*m_SecondsPerCount;
@@ -78,7 +78,6 @@ CClock::Process()
 
 	m_fTimeElapsed += m_fDeltaTime;
 }
-
 
 float
 CClock::GetDeltaTick()
